Update base on comments
